backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
tavily-python>=0.3.0
pydantic>=2.5.0
sse-starlette>=1.8.0
httpx>=0.25.0
aiohttp>=3.9.0
python-multipart>=0.0.9
pypdf>=4.0.0
groq>=0.3.0
Pillow>=10.0.0
